chore: moved physical planner attributes to their own file
